Add files via upload
This is Project on performing Exploratory Data Analysis (EDA) on "Super Market Data" and find out the business problems that the store faces. And also finding out the weak areas of profit so that they work in that areas to get profits. In this project I have done the EDA analysis using Python libraries like Numpy, Pandas and visualization libraries like Matplotlib, seaborn.